requirements.txt
asgiref==3.2.5
boto3==1.13.19
botocore==1.16.19
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
django-s3-storage==0.13.2
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
